% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{ana}
\alias{ana}
\title{The anamorphism, categorical dual of the catamorphism}
\usage{
ana(coalg, expr)
}
\arguments{
\item{coalg}{A coalgebra}

\item{expr}{A terminal expression}
}
\description{
The anamorphism, categorical dual of the catamorphism
}
