% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{apo}
\alias{apo}
\title{The apomorphism, categorical dual of the paramorphism}
\usage{
apo(coalg, expr)
}
\arguments{
\item{coalg}{A coalgebra}

\item{expr}{A terminal expression}
}
\description{
The apomorphism, categorical dual of the paramorphism
}
