% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-tuple.R
\name{as.character.tuple}
\alias{as.character.tuple}
\title{Get a text representation of a tuple}
\usage{
\method{as.character}{tuple}(t)
}
\arguments{
\item{t}{A tuple}
}
\description{
Get a text representation of a tuple
}
