% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-either.R
\name{bind.Either}
\alias{bind.Either}
\title{Bind a list}
\usage{
\method{bind}{Either}(ex, f)
}
\arguments{
\item{ex}{An Either monad}

\item{f}{A function to bind against the Either monad}
}
\description{
Bind a list
}
