% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free.R
\name{bind.free.monad}
\alias{bind.free.monad}
\title{Bind operation for a free monaad}
\usage{
\method{bind}{free.monad}(fr, f)
}
\arguments{
\item{fr}{A free monad}

\item{f}{A function to bind}
}
\description{
Bind operation for a free monaad
}
