% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{cata}
\alias{cata}
\title{The catamorphism}
\usage{
cata(alg, expr)
}
\arguments{
\item{alg}{An algebra}

\item{expr}{An initial expression}
}
\description{
The catamorphism
}
\examples{
len <- function (lst) {
  ln = function (st) {
    if (is.NilF (st)) {
      return (0)
    }
    else if (is.ConsF (st)) {
      return (1 + st$carrier)
    }
  }
  return (cata (ln, lst))
}
len (list (1,2,3,4))

}
