% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofree.R
\name{Cofree}
\alias{Cofree}
\title{Construct a cofree comonad}
\usage{
Cofree(attr, carrier)
}
\arguments{
\item{attr}{The comonad attribute}

\item{carrier}{The comonad carrier}
}
\description{
Construct a cofree comonad
}
