% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-base-list.R
\name{ConsF}
\alias{ConsF}
\title{Construct a \code{Cons} value of the list base functor}
\usage{
ConsF(x, xs)
}
\arguments{
\item{x}{First list element}

\item{xs}{Subsequent list elements, constructed with \code{Cons}}
}
\description{
Construct a \code{Cons} value of the list base functor
}
