% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-either.R
\name{either}
\alias{either}
\title{Run a function depending on whether an Either value is \code{Left} or \code{Right}}
\usage{
either(fl, fr, ex)
}
\arguments{
\item{fl}{A function to run on the value held by a \code{Left} value}

\item{fr}{A function to run on the value held by a \code{Right} value}

\item{ex}{A value to check}
}
\description{
Run a function depending on whether an Either value is \code{Left} or \code{Right}
}
