% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-base-list.R
\name{embed.ListF}
\alias{embed.ListF}
\title{Wrap a single layer of recursion}
\usage{
\method{embed}{ListF}(lstf)
}
\arguments{
\item{lstf}{A list base functor}
}
\description{
Wrap a single layer of recursion
}
