% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-either.R
\name{==.Either}
\alias{==.Either}
\title{Compare two Either monads}
\usage{
\method{==}{Either}(x, y)
}
\arguments{
\item{x}{An Either monad}

\item{y}{Another Either monad}
}
\description{
Compare two Either monads
}
