% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofree.R
\name{extend.cofree.comonad}
\alias{extend.cofree.comonad}
\title{Take a function and apply it to the comonad's attribute}
\usage{
\method{extend}{cofree.comonad}(cfr, f)
}
\arguments{
\item{cfr}{A cofree comonad}

\item{f}{A function to apply to the cofree comonad's attribute}
}
\description{
Take a function and apply it to the comonad's attribute
}
