% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-monad.R
\name{extend}
\alias{extend}
\title{Extend takes a function and applies it to the comonad's attribute}
\usage{
extend(cm, ...)
}
\arguments{
\item{cm}{A comonad}

\item{f}{A function}
}
\description{
Extend takes a function and applies it to the comonad's attribute
}
