% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofree.R
\name{extract.cofree.comonad}
\alias{extract.cofree.comonad}
\title{Extract a value from a cofree comonad}
\usage{
\method{extract}{cofree.comonad}(cfr)
}
\arguments{
\item{cfr}{A cofree comonad}
}
\description{
Extract a value from a cofree comonad
}
