% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-either.R
\name{fmap.Either}
\alias{fmap.Either}
\title{Map over an Either monad}
\usage{
\method{fmap}{Either}(ex, f)
}
\arguments{
\item{f}{A function to map}

\item{x}{An Either monad}
}
\description{
Map over an Either monad
}
