% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-list.R
\name{fmap.list}
\alias{fmap.list}
\title{Flat-map over a list}
\usage{
\method{fmap}{list}(X, FUN, ...)
}
\arguments{
\item{x}{A list}

\item{f}{A function to flat-map}
}
\description{
Flat-map over a list
}
