% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-functor.R
\name{fmap}
\alias{fmap}
\title{Flat-map over a functor}
\usage{
fmap(fnct, ...)
}
\arguments{
\item{x}{A functor}

\item{f}{A function to flat-map}
}
\description{
flat-map is implemented differently for each functor, with each functor
satisfying the functor laws.

Nonetheless this is set up so that unlike in Haskell, we use \code{fmap (x, f)}
rather than fmap f x. The reason for this is to enable use of \code{fmap()} in
native pipes without a placeholder, e.g. \verb{x |> fmap (\\(a) a+2)}.

In contrast, the infix operator \verb{\%<$>\%} works as in Haskell as this is more
natural.
}
