% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-tuple.R
\name{fmap.tuple}
\alias{fmap.tuple}
\title{Flat-map over a tuple}
\usage{
\method{fmap}{tuple}(t, f)
}
\arguments{
\item{t}{A tuple}

\item{f}{A function to flat-map}
}
\description{
Flat-map over a tuple
}
