% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{futu}
\alias{futu}
\title{The futumorphism, categorical dual of the histomorphism}
\usage{
futu(coalg, a)
}
\arguments{
\item{coalg}{A coalgebra}

\item{expr}{A terminal expression}
}
\description{
The futumorphism, categorical dual of the histomorphism
}
