% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofree.R
\name{\%:<\%}
\alias{\%:<\%}
\title{Infix constructor of a cofree comonad}
\usage{
attr \%:<\% carrier
}
\arguments{
\item{attr}{The comonad attribute}

\item{carrier}{The comonad carrier}
}
\description{
Infix constructor of a cofree comonad
}
