% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-base-list.R
\name{is.ConsF}
\alias{is.ConsF}
\title{Check if a value is a concatenated list base functor}
\usage{
is.ConsF(x)
}
\arguments{
\item{x}{The value to check}
}
\description{
Check if a value is a concatenated list base functor
}
