% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free.R
\name{is.Free}
\alias{is.Free}
\title{Check if the \code{Free} constructor governs an expression}
\usage{
is.Free(fr)
}
\arguments{
\item{fr}{A free monad}
}
\description{
Check if the \code{Free} constructor governs an expression
}
