% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-either.R
\name{is.Left}
\alias{is.Left}
\title{Check if a value is a \code{Right} Either monad}
\usage{
is.Left(x)
}
\arguments{
\item{x}{A value to check}
}
\description{
Check if a value is a \code{Right} Either monad
}
