% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free.R
\name{is.Pure}
\alias{is.Pure}
\title{Check if the \code{Pure} constructor governs an expression}
\usage{
is.Pure(fr)
}
\arguments{
\item{fr}{A free monad}
}
\description{
Check if the \code{Pure} constructor governs an expression
}
