% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-either.R
\name{Left}
\alias{Left}
\title{Construct a \code{Left} value of the Either monad}
\usage{
Left(x)
}
\arguments{
\item{x}{A value wrapped by \code{Left}}
}
\description{
Construct a \code{Left} value of the Either monad
}
