% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-monad.R
\name{mdiscard}
\alias{mdiscard}
\title{Bind a monad, albeit discard the initial result}
\usage{
mdiscard(m, f)
}
\arguments{
\item{m}{A monad}

\item{f}{A function to bind against the monad}
}
\description{
Bind a monad, albeit discard the initial result
}
