% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{para}
\alias{para}
\title{The paramorphism}
\usage{
para(alg, expr)
}
\arguments{
\item{alg}{An algebra}

\item{expr}{An initial expression}
}
\description{
The paramorphism
}
\examples{
tailL <- function (lst) {
    alg = function (l) {
        if (is.NilF (l))
            return (list())
        else {
            x  = l$attr
            xs = l$carrier
            xsa = first (xs)
            xsc = second (xs)

            return (xsa)
        }
    }
    return (para (alg, lst))
}

}
