% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-either.R
\name{Right}
\alias{Right}
\title{Construct a \code{Right} value of the Either monad}
\usage{
Right(x)
}
\arguments{
\item{x}{A value wrapped by \code{Right}}
}
\description{
Construct a \code{Right} value of the Either monad
}
