% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free.R
\name{fmap.free.monad}
\alias{fmap.free.monad}
\title{Map over a free monaad}
\usage{
\method{fmap}{free.monad}(fr, f)
}
\arguments{
\item{fr}{A free monad}

\item{f}{A function to map}
}
\description{
Map over a free monaad
}
