% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-base-list.R
\name{fmap.ListF}
\alias{fmap.ListF}
\title{Flat-map over a list base functor}
\usage{
\method{fmap}{ListF}(lstf, f)
}
\arguments{
\item{f}{A function to flat-map}

\item{x}{A list base functor}
}
\description{
Flat-map over a list base functor
}
